Move main files into `main`
