+++ src/main/SplashScreen.ts
