+++ src/main/TrayIcon.ts
